# medicare_backend/swagger/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "文档说明：《医疗保险报销系统》后端接口说明",
    "version": "v2.1.1",
    "title": "《医疗保险报销系统》后端接口说明",
    "contact": {
      "name": "skraua",
      "url": "https://github.com/skraua",
      "email": "ltghbwy@gamil.com"
    }
  },
  "host": "127.0.0.1:9999",
  "basePath": "/",
...
  "paths": {
...
    "/pdf/check/{fileName}": {
      "get": {
        "tags": ["PDF文件下载管理"],
        "summary": "检查PDF文件是否存在",
        "operationId": "checkPdfExistsUsingGET",
        "produces": ["*/*"],
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "description": "fileName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResultVo«string»",
              "originalRef": "ResultVo«string»"
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        },
        "deprecated": false
      }
    },
    "/pdf/clean-expired": {
      "post": {
        "tags": ["PDF文件下载管理"],
        "summary": "清理过期PDF文件",
        "operationId": "cleanExpiredFilesUsingPOST",
        "consumes": ["application/json"],
        "produces": ["*/*"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResultVo«string»",
              "originalRef": "ResultVo«string»"
            }
          },
          "201": { "description": "Created" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        },
        "deprecated": false
      }
    },
    "/pdf/download/{fileName}": {
      "get": {
        "tags": ["PDF文件下载管理"],
        "summary": "下载PDF文件",
        "operationId": "downloadPdfUsingGET",
        "produces": ["*/*"],
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "description": "fileName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Resource",
              "originalRef": "Resource"
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "File": {
      "type": "object",
      "properties": {
        "absolute": { "type": "boolean" },
        "absoluteFile": { "$ref": "#/definitions/File", "originalRef": "File" },
        "absolutePath": { "type": "string" },
        "canonicalFile": {
          "$ref": "#/definitions/File",
          "originalRef": "File"
        },
        "canonicalPath": { "type": "string" },
        "directory": { "type": "boolean" },
        "file": { "type": "boolean" },
        "freeSpace": { "type": "integer", "format": "int64" },
        "hidden": { "type": "boolean" },
        "name": { "type": "string" },
        "parent": { "type": "string" },
        "parentFile": { "$ref": "#/definitions/File", "originalRef": "File" },
        "path": { "type": "string" },
        "totalSpace": { "type": "integer", "format": "int64" },
        "usableSpace": { "type": "integer", "format": "int64" }
      },
      "title": "File"
    },
    "InputStream": { "type": "object", "title": "InputStream" },
...
    "Resource": {
      "type": "object",
      "properties": {
        "description": { "type": "string" },
        "file": { "$ref": "#/definitions/File", "originalRef": "File" },
        "filename": { "type": "string" },
        "inputStream": {
          "$ref": "#/definitions/InputStream",
          "originalRef": "InputStream"
        },
        "open": { "type": "boolean" },
        "readable": { "type": "boolean" },
        "uri": { "$ref": "#/definitions/URI", "originalRef": "URI" },
        "url": { "$ref": "#/definitions/URL", "originalRef": "URL" }
      },
      "title": "Resource"
    },
...
    "ResultVo«string»": {
      "type": "object",
      "properties": {
        "data": { "type": "string" },
        "flag": { "type": "boolean" },
        "message": { "type": "string" }
      },
      "title": "ResultVo«string»"
    },
...
    "URI": {
      "type": "object",
      "properties": {
        "absolute": { "type": "boolean" },
        "authority": { "type": "string" },
        "fragment": { "type": "string" },
        "host": { "type": "string" },
        "opaque": { "type": "boolean" },
        "path": { "type": "string" },
        "port": { "type": "integer", "format": "int32" },
        "query": { "type": "string" },
        "rawAuthority": { "type": "string" },
        "rawFragment": { "type": "string" },
        "rawPath": { "type": "string" },
        "rawQuery": { "type": "string" },
        "rawSchemeSpecificPart": { "type": "string" },
        "rawUserInfo": { "type": "string" },
        "scheme": { "type": "string" },
        "schemeSpecificPart": { "type": "string" },
        "userInfo": { "type": "string" }
      },
      "title": "URI"
    },
    "URL": {
      "type": "object",
      "properties": {
        "authority": { "type": "string" },
        "content": { "type": "object" },
        "defaultPort": { "type": "integer", "format": "int32" },
        "file": { "type": "string" },
        "host": { "type": "string" },
        "path": { "type": "string" },
        "port": { "type": "integer", "format": "int32" },
        "protocol": { "type": "string" },
        "query": { "type": "string" },
        "ref": { "type": "string" },
        "userInfo": { "type": "string" }
      },
      "title": "URL"
    },
...
  }
}